maj css + banner
Implantation des fichiers css pour chaque composant et chaque page
mise en place de la bannière
