+++ src/components/Banner/banner.css
@@ -1,0 +1,48 @@
+.banner {
+    width: 100%;
+    box-sizing: border-box;
+    position: relative;
+    display: flex;
+    max-width: 1240px;
+    height: 223px;
+    background: #000;
+    overflow: hidden;
+    flex-shrink: 0;
+    border-radius: 25px;
+    -webkit-border-radius: 25px;
+    -moz-border-radius: 25px;
+    -ms-border-radius: 25px;
+    -o-border-radius: 25px;
+}
+
+.banner p {
+    width: 100%;
+    text-align: center;
+    position: absolute;
+    align-self: center;
+    z-index: 10;
+    color: #FFF;
+    text-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+    -webkit-text-stroke-width: 1;
+    -webkit-text-stroke-color: #000;
+    font-family: "Montserrat", sans-serif;
+    font-size: 48px;
+    font-weight: 700;
+}
+
+.banner img {
+    width: 100%;
+    box-sizing: border-box;
+    object-fit: cover;
+    opacity: 0.6;
+    z-index: 0;
+}
+
+.container {
+    width: 100%;
+    max-width: 1240px;
+    height: fit-content;
+    display: flex;
+    flex-direction: column;
+    margin: auto;
+}
