--- src/components/Galery.jsx
@@ -1,11 +1,0 @@
-import Card from "./Card";
-
-const Gallery = () => {
-    return (
-        <div className="galery">
-            <Card/>
-        </div>
-    );
-};
-
-export default Gallery;
