+++ src/components/Gallery/Gallery.jsx
@@ -1,0 +1,11 @@
+import Card from "../Card/Card";
+
+const Gallery = () => {
+  return (
+    <div className="galery">
+      <Card />
+    </div>
+  );
+};
+
+export default Gallery;
