+++ src/components/Header/Header.jsx
@@ -1,4 +1,4 @@
-import NavBar from "./NavBar";
+import NavBar from "../NavBar/NavBar";
 
 const Header = () => {
     return (
