--- src/pages/Home.jsx
@@ -1,14 +1,0 @@
-import Banner from "../components/Banner/Banner";
-import Gallery from "../components/Galery";
-
-const Home = () => {
-    return (
-        <div>
-            <h2>Accueil</h2>
-            <Banner/>
-            <Gallery/>
-        </div>
-    );
-};
-
-export default Home;
