+++ src/pages/Home/Home.jsx
@@ -1,0 +1,14 @@
+import Banner from "../../components/Banner/Banner";
+import Gallery from "../../components/Gallery/Gallery";
+
+const Home = () => {
+  return (
+    <div className="container">
+      <h2>Accueil</h2>
+      <Banner />
+      <Gallery />
+    </div>
+  );
+};
+
+export default Home;
